app/server/requirements.txt
# === Core ===
python-dotenv==1.0.1
requests==2.32.3
uuid==1.30
tqdm==4.66.5
numpy==1.26.4

# === Mastodon Bot ===
Mastodon.py==1.8.1

# === LLM + AI ===
google-generativeai==0.8.3
openai==1.52.0
sentence-transformers==3.0.1

# === LlamaIndex (RAG pipeline) ===
llama-index-core==0.14.6
llama-index-embeddings-huggingface==0.6.1
llama-index-vector-stores-chroma==0.5.3
chromadb==1.2.1

# === OCR + PDF Parsing ===
pymupdf==1.22.5
pytesseract==0.3.13
pillow==10.4.0


# === Web Server ===
fastapi
uvicorn
beanie
motor
strawberry-graphql
pydantic
python-dotenv
